% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/provenance_dag.R
\name{add_provenance_node}
\alias{add_provenance_node}
\alias{add_provenance_edge}
\title{Add a node or edge to a provenance DAG}
\usage{
add_provenance_node(dag, node)

add_provenance_edge(dag, edge)
}
\arguments{
\item{dag}{A provenance DAG.}

\item{node, edge}{Object to add.}
}
\value{
An updated validated DAG.
}
\description{
Add a node or edge to a provenance DAG
}
