% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_remaining_registry.R
\name{admixture_module_spec}
\alias{admixture_module_spec}
\title{Construct the ADMIXTURE module descriptor}
\usage{
admixture_module_spec()
}
\value{
A `PopgenVCFModuleSpec` object.
}
\description{
Construct the ADMIXTURE module descriptor
}
