% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ancestry_consensus.R
\name{ancestry_consensus_table}
\alias{ancestry_consensus_table}
\title{Convert consensus ancestry estimates to long form}
\usage{
ancestry_consensus_table(x)
}
\arguments{
\item{x}{A `PopgenVCFAncestryConsensus` object.}
}
\value{
A `data.table` with mean, median, variance, standard deviation, and
  confidence interval values for every sample and cluster.
}
\description{
Convert consensus ancestry estimates to long form
}
