% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ancestry_k_selection.R
\name{ancestry_k_selection_text}
\alias{ancestry_k_selection_text}
\title{Generate manuscript-ready ancestry K-selection text}
\usage{
ancestry_k_selection_text(x)
}
\arguments{
\item{x}{A `PopgenVCFKSelection` object.}
}
\value{
Named character vector containing methods and results text.
}
\description{
Generate manuscript-ready ancestry K-selection text
}
