% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ancestry_result.R
\name{ancestry_q_table}
\alias{ancestry_q_table}
\title{Convert an ancestry Q matrix to long-form source data}
\usage{
ancestry_q_table(x)
}
\arguments{
\item{x}{A `PopgenVCFAncestryReplicate` object.}
}
\value{
A `data.table` with sample, cluster, and ancestry coefficient.
}
\description{
Convert an ancestry Q matrix to long-form source data
}
