% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/artifacts.R
\name{artifact_manifest_table}
\alias{artifact_manifest_table}
\title{Convert an artifact manifest to a data table}
\usage{
artifact_manifest_table(manifest)
}
\arguments{
\item{manifest}{A `PopgenVCFArtifactManifest`.}
}
\value{
A data table with one row per artifact.
}
\description{
Convert an artifact manifest to a data table
}
