% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regression_archive.R
\name{benchmark_archive_table}
\alias{benchmark_archive_table}
\title{Convert archive objects to stable tables}
\usage{
benchmark_archive_table(x)
}
\arguments{
\item{x}{A release record or benchmark archive.}
}
\value{
A data table.
}
\description{
Convert archive objects to stable tables
}
