% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benchmark.R
\name{benchmark_stage}
\alias{benchmark_stage}
\title{Benchmark a popgenVCF expression}
\usage{
benchmark_stage(label, expr)
}
\arguments{
\item{label}{Benchmark label.}

\item{expr}{Expression to evaluate.}
}
\value{
One-row data table with elapsed time and garbage-collection change.
}
\description{
Benchmark a popgenVCF expression
}
