% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benchmark_registry.R
\name{benchmark_suite_table}
\alias{benchmark_suite_table}
\title{Convert a benchmark suite to a summary table}
\usage{
benchmark_suite_table(x)
}
\arguments{
\item{x}{A benchmark suite.}
}
\value{
A data table.
}
\description{
Convert a benchmark suite to a summary table
}
