% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dashboard.R
\name{calculate_scientific_quality}
\alias{calculate_scientific_quality}
\title{Calculate a transparent scientific quality score}
\usage{
calculate_scientific_quality(plan)
}
\arguments{
\item{plan}{A `PopgenVCFReportPlan`.}
}
\value{
A list containing component scores and an overall percentage.
}
\description{
Calculate a transparent scientific quality score
}
