% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/canonical_1000g_chrY.R
\name{canonical_1000g_chrY_source}
\alias{canonical_1000g_chrY_source}
\title{Describe the approved 1000 Genomes Phase 3 chromosome Y source}
\usage{
canonical_1000g_chrY_source()
}
\value{
A named source specification.
}
\description{
Describe the approved 1000 Genomes Phase 3 chromosome Y source
}
