% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/canonical_drift_detection.R
\name{canonical_drift_table}
\alias{canonical_drift_table}
\title{Return a canonical drift assessment table}
\usage{
canonical_drift_table(assessment)
}
\arguments{
\item{assessment}{Canonical drift assessment.}
}
\value{
Deterministically ordered data frame.
}
\description{
Return a canonical drift assessment table
}
