% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_remaining_registry.R
\name{chromosome_module_spec}
\alias{chromosome_module_spec}
\title{Construct the chromosome-specific analysis module descriptor}
\usage{
chromosome_module_spec()
}
\value{
A `PopgenVCFModuleSpec` object.
}
\description{
Construct the chromosome-specific analysis module descriptor
}
