% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_results.R
\name{core_result_table}
\alias{core_result_table}
\title{Extract the primary tabular representation of a core result}
\usage{
core_result_table(x)
}
\arguments{
\item{x}{A `PopgenVCFCoreResult`.}
}
\value{
A data table.
}
\description{
Extract the primary tabular representation of a core result
}
