% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{default_config}
\alias{default_config}
\title{Create the default popgenVCF configuration}
\usage{
default_config()
}
\value{
A nested configuration list using the supported schema.
}
\description{
Create the default popgenVCF configuration
}
