% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_remaining_registry.R
\name{diversity_module_spec}
\alias{diversity_module_spec}
\title{Construct the diversity analysis module descriptor}
\usage{
diversity_module_spec()
}
\value{
A `PopgenVCFModuleSpec` object.
}
\description{
Construct the diversity analysis module descriptor
}
