% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/external_reference_validation.R
\name{external_reference_table}
\alias{external_reference_table}
\title{Convert external-reference results to a long-form table}
\usage{
external_reference_table(results)
}
\arguments{
\item{results}{A result or list of results.}
}
\value{
A data table.
}
\description{
Convert external-reference results to a long-form table
}
