% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fair_export.R
\name{fair_identifier}
\alias{fair_identifier}
\title{Return stable FAIR identifiers}
\usage{
fair_identifier(project, artifact_id = NULL)
}
\arguments{
\item{project}{A `PopgenVCFProject`.}

\item{artifact_id}{Optional artifact identifier.}
}
\value{
A stable URN.
}
\description{
Return stable FAIR identifiers
}
