% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_remaining_registry.R
\name{faststructure_module_spec}
\alias{faststructure_module_spec}
\title{Construct the fastStructure module descriptor}
\usage{
faststructure_module_spec()
}
\value{
A `PopgenVCFModuleSpec` object.
}
\description{
Construct the fastStructure module descriptor
}
