% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzzzzzzz_journal_submission_profiles.R
\name{generic_journal_profile}
\alias{generic_journal_profile}
\title{Return the generic journal profile}
\usage{
generic_journal_profile()
}
\value{
A `PopgenVCFJournalProfile`.
}
\description{
Return the generic journal profile
}
