% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzzzzzzzz_reporting_checklists.R
\name{generic_reporting_checklist}
\alias{generic_reporting_checklist}
\title{Return the generic population-genomics reporting checklist}
\usage{
generic_reporting_checklist()
}
\value{
A `PopgenVCFReportingChecklist`.
}
\description{
Return the generic population-genomics reporting checklist
}
