% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regression_archive.R
\name{get_release_benchmark}
\alias{get_release_benchmark}
\title{Retrieve a release from an archive}
\usage{
get_release_benchmark(archive, release)
}
\arguments{
\item{archive}{A benchmark archive.}

\item{release}{Release identifier.}
}
\value{
A release benchmark record.
}
\description{
Retrieve a release from an archive
}
