% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/golden_outputs.R
\name{golden_output_table}
\alias{golden_output_table}
\title{Convert golden objects to stable tables}
\usage{
golden_output_table(x)
}
\arguments{
\item{x}{A golden store or result.}
}
\value{
A data table.
}
\description{
Convert golden objects to stable tables
}
