% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_object.R
\name{is_popgen_vcf_analysis}
\alias{is_popgen_vcf_analysis}
\title{Test whether an object is a popgenVCF analysis state}
\usage{
is_popgen_vcf_analysis(x)
}
\arguments{
\item{x}{Any R object.}
}
\value{
A logical scalar.
}
\description{
Test whether an object is a popgenVCF analysis state
}
