% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzzzzzzz_journal_submission_profiles.R
\name{journal_profile}
\alias{journal_profile}
\title{Return a built-in generic journal profile}
\usage{
journal_profile(
  name = c("research-article", "short-communication", "data-note")
)
}
\arguments{
\item{name}{One of `research-article`, `short-communication`, or `data-note`.}
}
\value{
A `PopgenVCFJournalProfile`.
}
\description{
Return a built-in generic journal profile
}
