% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/artifact_lineage.R
\name{lineage_execution_table}
\alias{lineage_execution_table}
\alias{lineage_artifact_table}
\title{Convert lineage records to stable tables}
\usage{
lineage_execution_table(lineage)

lineage_artifact_table(lineage)
}
\arguments{
\item{lineage}{An artifact lineage object.}
}
\value{
A data table.
}
\description{
Convert lineage records to stable tables
}
