% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/registry.R
\name{list_analyses}
\alias{list_analyses}
\title{List registered analysis modules}
\usage{
list_analyses(registry)
}
\arguments{
\item{registry}{A `PopgenVCFRegistry` object.}
}
\value{
A data table describing registered modules.
}
\description{
List registered analysis modules
}
