% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benchmark_registry.R
\name{list_benchmarks}
\alias{list_benchmarks}
\title{List registered benchmarks}
\usage{
list_benchmarks(registry)
}
\arguments{
\item{registry}{A benchmark registry.}
}
\value{
A data table.
}
\description{
List registered benchmarks
}
