% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzzzzzzz_manuscript_regeneration_plans.R
\name{manuscript_regeneration_table}
\alias{manuscript_regeneration_table}
\title{Compute the deterministic regeneration table}
\usage{
manuscript_regeneration_table(x)
}
\arguments{
\item{x}{A regeneration plan or plan-like list.}
}
\value{
A deterministic data table.
}
\description{
Compute the deterministic regeneration table
}
