% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ancestry_backend.R
\name{new_ancestry_backend_registry}
\alias{new_ancestry_backend_registry}
\title{Create an ancestry backend registry}
\usage{
new_ancestry_backend_registry(backends = list())
}
\arguments{
\item{backends}{Optional list of backend specifications.}
}
\value{
A `PopgenVCFAncestryBackendRegistry`.
}
\description{
Create an ancestry backend registry
}
