% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ancestry_result.R
\name{new_ancestry_result}
\alias{new_ancestry_result}
\title{Create an ancestry result collection}
\usage{
new_ancestry_result(replicates)
}
\arguments{
\item{replicates}{List of `PopgenVCFAncestryReplicate` objects.}
}
\value{
A validated `PopgenVCFAncestryResult` object.
}
\description{
Create an ancestry result collection
}
