% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/artifact_lineage.R
\name{new_artifact_lineage}
\alias{new_artifact_lineage}
\title{Build immutable artifact lineage}
\usage{
new_artifact_lineage(executions = list(), artifacts = list())
}
\arguments{
\item{executions}{List of execution records.}

\item{artifacts}{List of artifact records.}
}
\value{
A validated `PopgenVCFArtifactLineage`.
}
\description{
Build immutable artifact lineage
}
