% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/artifacts.R
\name{new_artifact_manifest}
\alias{new_artifact_manifest}
\title{Create an artifact manifest}
\usage{
new_artifact_manifest(artifacts = list())
}
\arguments{
\item{artifacts}{A list of `PopgenVCFArtifact` objects.}
}
\value{
A `PopgenVCFArtifactManifest` object.
}
\description{
Create an artifact manifest
}
