% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benchmark_datasets.R
\name{new_benchmark_dataset_catalogue}
\alias{new_benchmark_dataset_catalogue}
\alias{register_benchmark_dataset}
\title{Create and modify a benchmark dataset catalogue}
\usage{
new_benchmark_dataset_catalogue(entries = list())

register_benchmark_dataset(catalogue, entry)
}
\arguments{
\item{entries}{Optional list of entries.}

\item{catalogue}{A benchmark dataset catalogue.}

\item{entry}{A dataset entry.}
}
\value{
A `PopgenVCFBenchmarkDatasetCatalogue`.
}
\description{
Create and modify a benchmark dataset catalogue
}
