% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benchmark_registry.R
\name{new_benchmark_registry}
\alias{new_benchmark_registry}
\title{Create a benchmark registry}
\usage{
new_benchmark_registry(specs = list())
}
\arguments{
\item{specs}{Optional list of benchmark specifications.}
}
\value{
A `PopgenVCFBenchmarkRegistry`.
}
\description{
Create a benchmark registry
}
