% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/canonical-scientific-objects.R
\name{new_canonical_distance}
\alias{new_canonical_distance}
\title{Construct a canonical genetic-distance object}
\usage{
new_canonical_distance(
  distance,
  entity_ids,
  metric,
  units = "unitless",
  provenance = list()
)
}
\description{
Construct a canonical genetic-distance object
}
\keyword{internal}
