% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/canonical-scientific-objects.R
\name{new_canonical_frequency}
\alias{new_canonical_frequency}
\title{Construct a canonical allele-frequency object}
\usage{
new_canonical_frequency(
  frequencies,
  population_ids,
  variant_ids,
  allele,
  provenance = list()
)
}
\description{
Construct a canonical allele-frequency object
}
\keyword{internal}
