% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/canonical-scientific-objects.R
\name{new_canonical_genotype}
\alias{new_canonical_genotype}
\title{Construct a canonical genotype object}
\usage{
new_canonical_genotype(
  genotypes,
  sample_ids,
  variant_ids,
  ploidy = 2L,
  allele_coding,
  genome_build = NULL,
  provenance = list()
)
}
\description{
Construct a canonical genotype object
}
\keyword{internal}
