% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/canonical-scientific-objects.R
\name{new_canonical_metadata}
\alias{new_canonical_metadata}
\title{Construct canonical sample/population metadata}
\usage{
new_canonical_metadata(
  data,
  sample_id_column = "sample_id",
  population_column = NULL,
  provenance = list()
)
}
\description{
Construct canonical sample/population metadata
}
\keyword{internal}
