% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/execution-resource-policy.R
\name{new_module_resource_requirements}
\alias{new_module_resource_requirements}
\title{Create module resource requirements}
\usage{
new_module_resource_requirements(
  threads = 1L,
  memory_mb = 0,
  temp_mb = 0,
  processes = 1L
)
}
\arguments{
\item{threads}{Required threads.}

\item{memory_mb}{Required memory in MiB.}

\item{temp_mb}{Required temporary storage in MiB.}

\item{processes}{Required processes.}
}
\value{
A validated named numeric vector.
}
\description{
Create module resource requirements
}
