% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/provenance_dag.R
\name{new_provenance_dag}
\alias{new_provenance_dag}
\title{Create and validate a provenance DAG}
\usage{
new_provenance_dag(nodes = list(), edges = list())
}
\arguments{
\item{nodes}{List of provenance nodes.}

\item{edges}{List of provenance edges.}
}
\value{
A validated `PopgenVCFProvenanceDAG`.
}
\description{
Create and validate a provenance DAG
}
