% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runtime-migration.R
\name{new_runtime_migration_registry}
\alias{new_runtime_migration_registry}
\title{Create an empty runtime migration registry}
\usage{
new_runtime_migration_registry()
}
\value{
A `PopgenVCFRuntimeMigrationRegistry` object.
}
\description{
Create an empty runtime migration registry
}
