% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/performance_benchmarks.R
\name{performance_benchmark_table}
\alias{performance_benchmark_table}
\title{Convert performance objects to stable tables}
\usage{
performance_benchmark_table(x)
}
\arguments{
\item{x}{A performance result or comparison.}
}
\value{
A data table.
}
\description{
Convert performance objects to stable tables
}
