% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/external_process_async.R
\name{poll_supervised_external_command}
\alias{poll_supervised_external_command}
\title{Poll an asynchronous supervised external command}
\usage{
poll_supervised_external_command(handle, timeout_milliseconds = 0)
}
\arguments{
\item{handle}{A `PopgenVCFExternalProcessHandle`.}

\item{timeout_milliseconds}{Maximum time to wait for process I/O during this
poll operation.}
}
\value{
`handle`, invisibly, after updating lifecycle state and output.
}
\description{
Poll an asynchronous supervised external command
}
