\name{popgenVCF-package}
\alias{popgenVCF}
\alias{popgenVCF-package}
\title{Population Genomics Toolkit for VCF Data}
\description{
A modular command-line toolkit and R API for reproducible analysis of diploid,
biallelic SNP VCF data, including quality control, population structure, genetic
diversity, differentiation, spatial genetics, chromosome-specific analyses,
external ADMIXTURE integration, and manuscript-oriented reporting.
}
\docType{package}
\author{Marc-Olivier Duceppe}
\keyword{package}
\keyword{genetics}
