% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/execution-cancellation.R
\name{print.PopgenVCFExecutionCancellationToken}
\alias{print.PopgenVCFExecutionCancellationToken}
\title{Print an execution cancellation token}
\usage{
\method{print}{PopgenVCFExecutionCancellationToken}(x, ...)
}
\arguments{
\item{x}{A cancellation token.}

\item{...}{Unused.}
}
\value{
`x`, invisibly.
}
\description{
Print an execution cancellation token
}
