% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/execution-checkpoint-fixes.R
\name{print.PopgenVCFExecutionCheckpoint}
\alias{print.PopgenVCFExecutionCheckpoint}
\title{Print an execution checkpoint}
\usage{
\method{print}{PopgenVCFExecutionCheckpoint}(x, ...)
}
\arguments{
\item{x}{A `PopgenVCFExecutionCheckpoint` object.}

\item{...}{Unused.}
}
\value{
`x`, invisibly.
}
\description{
Print an execution checkpoint
}
