% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/execution-engine.R
\name{print.PopgenVCFExecutionPlan}
\alias{print.PopgenVCFExecutionPlan}
\title{Print an analysis execution plan}
\usage{
\method{print}{PopgenVCFExecutionPlan}(x, ...)
}
\arguments{
\item{x}{A `PopgenVCFExecutionPlan` object.}

\item{...}{Unused.}
}
\value{
`x`, invisibly.
}
\description{
Print an analysis execution plan
}
