% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/execution-retry.R
\name{print.PopgenVCFExecutionRetryPolicy}
\alias{print.PopgenVCFExecutionRetryPolicy}
\title{Print an execution retry policy}
\usage{
\method{print}{PopgenVCFExecutionRetryPolicy}(x, ...)
}
\arguments{
\item{x}{A `PopgenVCFExecutionRetryPolicy` object.}

\item{...}{Unused.}
}
\value{
`x`, invisibly.
}
\description{
Print an execution retry policy
}
