% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/execution-timeout.R
\name{print.PopgenVCFExecutionTimeoutPolicy}
\alias{print.PopgenVCFExecutionTimeoutPolicy}
\title{Print an execution timeout policy}
\usage{
\method{print}{PopgenVCFExecutionTimeoutPolicy}(x, ...)
}
\arguments{
\item{x}{A `PopgenVCFExecutionTimeoutPolicy` object.}

\item{...}{Unused.}
}
\value{
`x`, invisibly.
}
\description{
Print an execution timeout policy
}
