% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/external-process.R
\name{print.PopgenVCFExternalCommand}
\alias{print.PopgenVCFExternalCommand}
\title{Print an external command specification}
\usage{
\method{print}{PopgenVCFExternalCommand}(x, ...)
}
\arguments{
\item{x}{A `PopgenVCFExternalCommand` object.}

\item{...}{Unused.}
}
\value{
`x`, invisibly.
}
\description{
Print an external command specification
}
