% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/external_process_async.R
\name{print.PopgenVCFExternalProcessHandle}
\alias{print.PopgenVCFExternalProcessHandle}
\title{Print an asynchronous external-process handle}
\usage{
\method{print}{PopgenVCFExternalProcessHandle}(x, ...)
}
\arguments{
\item{x}{A process handle.}

\item{...}{Unused.}
}
\value{
`x`, invisibly.
}
\description{
Print an asynchronous external-process handle
}
