% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/external-process.R
\name{print.PopgenVCFExternalProcessResult}
\alias{print.PopgenVCFExternalProcessResult}
\title{Print an external process result}
\usage{
\method{print}{PopgenVCFExternalProcessResult}(x, ...)
}
\arguments{
\item{x}{A `PopgenVCFExternalProcessResult` object.}

\item{...}{Unused.}
}
\value{
`x`, invisibly.
}
\description{
Print an external process result
}
