% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projects.R
\name{project_table}
\alias{project_table}
\title{Convert project metadata to stable tables}
\usage{
project_table(x)
}
\arguments{
\item{x}{A `PopgenVCFProject` or `PopgenVCFProjectComparison`.}
}
\value{
A data table.
}
\description{
Convert project metadata to stable tables
}
