% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/provenance_dag.R
\name{provenance_ancestors}
\alias{provenance_ancestors}
\alias{provenance_descendants}
\title{Trace provenance ancestors or descendants}
\usage{
provenance_ancestors(dag, id)

provenance_descendants(dag, id)
}
\arguments{
\item{dag}{A provenance DAG.}

\item{id}{Node identifier.}
}
\value{
Character node identifiers.
}
\description{
Trace provenance ancestors or descendants
}
