% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/provenance_dag.R
\name{provenance_node_table}
\alias{provenance_node_table}
\alias{provenance_edge_table}
\title{Convert provenance objects to stable tables}
\usage{
provenance_node_table(dag)

provenance_edge_table(dag)
}
\arguments{
\item{dag}{A provenance DAG.}
}
\value{
A data table.
}
\description{
Convert provenance objects to stable tables
}
