% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/provenance_dag.R
\name{provenance_topological_order}
\alias{provenance_topological_order}
\title{Return deterministic topological order}
\usage{
provenance_topological_order(dag)
}
\arguments{
\item{dag}{A provenance DAG.}
}
\value{
Character node identifiers.
}
\description{
Return deterministic topological order
}
