% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_publication.R
\name{publication_analysis_narratives}
\alias{publication_analysis_narratives}
\title{Build analysis-specific publication narratives}
\usage{
publication_analysis_narratives(project)
}
\arguments{
\item{project}{A reproducible project.}
}
\value{
A data table with methods, legends, and citation keys.
}
\description{
Build analysis-specific publication narratives
}
