% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{read_config}
\alias{read_config}
\title{Read and merge a popgenVCF configuration}
\usage{
read_config(path)
}
\arguments{
\item{path}{YAML configuration file.}
}
\value{
The user configuration merged with current defaults.
}
\description{
Read and merge a popgenVCF configuration
}
