% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/external_process_async.R
\name{read_supervised_external_output}
\alias{read_supervised_external_output}
\title{Read accumulated asynchronous process output}
\usage{
read_supervised_external_output(handle)
}
\arguments{
\item{handle}{A `PopgenVCFExternalProcessHandle`.}
}
\value{
A list containing accumulated `stdout` and `stderr` strings.
}
\description{
Read accumulated asynchronous process output
}
