% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_spec.R
\name{register_analysis_module}
\alias{register_analysis_module}
\title{Register a self-contained analysis module descriptor}
\usage{
register_analysis_module(registry, module)
}
\arguments{
\item{registry}{A `PopgenVCFRegistry` object.}

\item{module}{A `PopgenVCFModuleSpec` object.}
}
\value{
The updated registry.
}
\description{
Register a self-contained analysis module descriptor
}
