% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/artifacts.R
\name{register_artifact}
\alias{register_artifact}
\title{Add an artifact to a manifest}
\usage{
register_artifact(manifest, artifact)
}
\arguments{
\item{manifest}{A `PopgenVCFArtifactManifest`.}

\item{artifact}{A `PopgenVCFArtifact`.}
}
\value{
Updated manifest.
}
\description{
Add an artifact to a manifest
}
