% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benchmark_registry.R
\name{register_benchmark}
\alias{register_benchmark}
\title{Register a benchmark}
\usage{
register_benchmark(registry, spec)
}
\arguments{
\item{registry}{A benchmark registry.}

\item{spec}{A benchmark specification.}
}
\value{
Updated registry.
}
\description{
Register a benchmark
}
