% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regression_archive.R
\name{register_release_benchmark}
\alias{register_release_benchmark}
\title{Register a release in an archive}
\usage{
register_release_benchmark(archive, record)
}
\arguments{
\item{archive}{A benchmark archive.}

\item{record}{A release benchmark record.}
}
\value{
Updated archive.
}
\description{
Register a release in an archive
}
