% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schema-validation.R
\name{register_schema}
\alias{register_schema}
\title{Register a schema descriptor}
\usage{
register_schema(registry, schema)
}
\arguments{
\item{registry}{Schema registry.}

\item{schema}{Schema descriptor.}
}
\value{
Updated schema registry.
}
\description{
Register a schema descriptor
}
