% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regression_reporting.R
\name{release_comparison_table}
\alias{release_comparison_table}
\title{Convert a release comparison to a stable table}
\usage{
release_comparison_table(x)
}
\arguments{
\item{x}{A `PopgenVCFReleaseComparison`.}
}
\value{
A data table.
}
\description{
Convert a release comparison to a stable table
}
