% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzzzzz_manuscript_jats.R
\name{render_manuscript_jats}
\alias{render_manuscript_jats}
\title{Render deterministic JATS XML}
\usage{
render_manuscript_jats(manuscript)
}
\arguments{
\item{manuscript}{A validated `PopgenVCFManuscript`.}
}
\value{
A character vector containing JATS XML.
}
\description{
Render deterministic JATS XML
}
