% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzzzzzzz_manuscript_regeneration_plans.R
\name{render_manuscript_regeneration_plan}
\alias{render_manuscript_regeneration_plan}
\title{Render a manuscript regeneration plan as Markdown}
\usage{
render_manuscript_regeneration_plan(x)
}
\arguments{
\item{x}{A validated regeneration plan.}
}
\value{
Markdown lines.
}
\description{
Render a manuscript regeneration plan as Markdown
}
