% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manuscript_revision.R
\name{render_manuscript_revision_diff}
\alias{render_manuscript_revision_diff}
\title{Render a manuscript revision diff as Markdown}
\usage{
render_manuscript_revision_diff(diff)
}
\arguments{
\item{diff}{A `PopgenVCFManuscriptRevisionDiff`.}
}
\value{
Character vector containing Markdown.
}
\description{
Render a manuscript revision diff as Markdown
}
