% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reviewer_response.R
\name{reviewer_response_table}
\alias{reviewer_response_table}
\title{Return reviewer responses as a deterministic table}
\usage{
reviewer_response_table(x)
}
\arguments{
\item{x}{A validated reviewer-response record.}
}
\value{
A data table.
}
\description{
Return reviewer responses as a deterministic table
}
