% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benchmark_registry.R
\name{run_benchmark}
\alias{run_benchmark}
\title{Execute one benchmark specification}
\usage{
run_benchmark(spec)
}
\arguments{
\item{spec}{A benchmark specification.}
}
\value{
A `PopgenVCFBenchmarkResult`.
}
\description{
Execute one benchmark specification
}
