% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/performance_benchmarks.R
\name{run_performance_benchmark}
\alias{run_performance_benchmark}
\title{Run a performance benchmark}
\usage{
run_performance_benchmark(
  spec,
  fingerprint = performance_environment_fingerprint()
)
}
\arguments{
\item{spec}{A `PopgenVCFPerformanceSpec`.}

\item{fingerprint}{Optional environment fingerprint.}
}
\value{
A `PopgenVCFPerformanceResult`.
}
\description{
Run a performance benchmark
}
