% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_object.R
\name{set_analysis_result}
\alias{set_analysis_result}
\title{Store a result in an analysis state}
\usage{
set_analysis_result(x, name, value)
}
\arguments{
\item{x}{A `PopgenVCFAnalysis` object.}

\item{name}{Result name.}

\item{value}{Result value.}
}
\value{
Updated `PopgenVCFAnalysis` object.
}
\description{
Store a result in an analysis state
}
