% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzzz_manuscript_csl_integration.R
\name{set_manuscript_citation_profile}
\alias{set_manuscript_citation_profile}
\title{Attach a citation profile to a manuscript}
\usage{
set_manuscript_citation_profile(manuscript, profile)
}
\arguments{
\item{manuscript}{A validated `PopgenVCFManuscript`.}

\item{profile}{A validated `PopgenVCFCitationProfile`.}
}
\value{
An updated manuscript object.
}
\description{
Attach a citation profile to a manuscript
}
