% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fair_export.R
\name{set_project_fair_metadata}
\alias{set_project_fair_metadata}
\title{Attach FAIR metadata to a project}
\usage{
set_project_fair_metadata(project, metadata)
}
\arguments{
\item{project}{A reproducible project.}

\item{metadata}{FAIR metadata for the same project.}
}
\value{
Updated project.
}
\description{
Attach FAIR metadata to a project
}
