% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/publication_companion.R
\name{set_project_publication_bundle}
\alias{set_project_publication_bundle}
\title{Attach a publication plan to a project}
\usage{
set_project_publication_bundle(project, bundle)
}
\arguments{
\item{project}{A reproducible project.}

\item{bundle}{A `PopgenVCFPublicationBundle`.}
}
\value{
Updated project.
}
\description{
Attach a publication plan to a project
}
