% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_remaining_registry.R
\name{snmf_module_spec}
\alias{snmf_module_spec}
\title{Construct the sNMF module descriptor}
\usage{
snmf_module_spec()
}
\value{
A `PopgenVCFModuleSpec` object.
}
\description{
Construct the sNMF module descriptor
}
