% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_remaining_registry.R
\name{tree_module_spec}
\alias{tree_module_spec}
\title{Construct the neighbour-joining tree module descriptor}
\usage{
tree_module_spec()
}
\value{
A `PopgenVCFModuleSpec` object.
}
\description{
Construct the neighbour-joining tree module descriptor
}
