% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ancestry_result.R
\name{validate_ancestry_result}
\alias{validate_ancestry_result}
\title{Validate an ancestry result collection}
\usage{
validate_ancestry_result(x)
}
\arguments{
\item{x}{A `PopgenVCFAncestryResult` object.}
}
\value{
`x`, invisibly, when valid.
}
\description{
Validate an ancestry result collection
}
