% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/artifact_lineage.R
\name{validate_artifact_lineage}
\alias{validate_artifact_lineage}
\title{Validate immutable artifact lineage}
\usage{
validate_artifact_lineage(lineage)
}
\arguments{
\item{lineage}{A `PopgenVCFArtifactLineage`.}
}
\value{
`lineage`, invisibly.
}
\description{
Validate immutable artifact lineage
}
