% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/artifacts.R
\name{validate_artifact_manifest}
\alias{validate_artifact_manifest}
\title{Validate an artifact manifest}
\usage{
validate_artifact_manifest(manifest, must_exist = FALSE)
}
\arguments{
\item{manifest}{A `PopgenVCFArtifactManifest`.}

\item{must_exist}{Require all required artifact files to exist.}
}
\value{
`TRUE` invisibly, or an error.
}
\description{
Validate an artifact manifest
}
