% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benchmark_registry.R
\name{validate_benchmark_dataset}
\alias{validate_benchmark_dataset}
\title{Validate a benchmark dataset}
\usage{
validate_benchmark_dataset(x)
}
\arguments{
\item{x}{A `PopgenVCFBenchmarkDataset`.}
}
\value{
`x`, invisibly.
}
\description{
Validate a benchmark dataset
}
