% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benchmark_datasets.R
\name{validate_benchmark_dataset_entry}
\alias{validate_benchmark_dataset_entry}
\title{Validate a benchmark dataset catalogue entry}
\usage{
validate_benchmark_dataset_entry(x)
}
\arguments{
\item{x}{A `PopgenVCFBenchmarkDatasetEntry`.}
}
\value{
`x`, invisibly.
}
\description{
Validate a benchmark dataset catalogue entry
}
