% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benchmark_registry.R
\name{validate_benchmark_spec}
\alias{validate_benchmark_spec}
\title{Validate a benchmark specification}
\usage{
validate_benchmark_spec(x)
}
\arguments{
\item{x}{A `PopgenVCFBenchmarkSpec`.}
}
\value{
`x`, invisibly.
}
\description{
Validate a benchmark specification
}
