% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benchmark_registry.R
\name{validate_benchmark_suite}
\alias{validate_benchmark_suite}
\title{Validate a benchmark suite}
\usage{
validate_benchmark_suite(x)
}
\arguments{
\item{x}{A `PopgenVCFBenchmarkSuite`.}
}
\value{
`x`, invisibly.
}
\description{
Validate a benchmark suite
}
