% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/canonical_baseline_metrics.R
\name{validate_canonical_baseline_metric}
\alias{validate_canonical_baseline_metric}
\title{Validate a canonical baseline metric}
\usage{
validate_canonical_baseline_metric(metric)
}
\arguments{
\item{metric}{Baseline metric.}
}
\value{
`metric`, invisibly.
}
\description{
Validate a canonical baseline metric
}
