% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/canonical_dataset_registry.R
\name{validate_canonical_dataset_registry}
\alias{validate_canonical_dataset_registry}
\title{Validate a canonical dataset registry}
\usage{
validate_canonical_dataset_registry(registry)
}
\arguments{
\item{registry}{Canonical dataset registry.}
}
\value{
`registry`, invisibly.
}
\description{
Validate a canonical dataset registry
}
