% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/canonical_1000g_chrY.R
\name{validate_canonical_source}
\alias{validate_canonical_source}
\title{Validate an approved canonical source specification}
\usage{
validate_canonical_source(source)
}
\arguments{
\item{source}{Canonical source specification.}
}
\value{
`source`, invisibly.
}
\description{
Validate an approved canonical source specification
}
