% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/canonical_validation_dataset.R
\name{validate_canonical_validation_dataset}
\alias{validate_canonical_validation_dataset}
\title{Validate a canonical validation dataset specification}
\usage{
validate_canonical_validation_dataset(spec)
}
\arguments{
\item{spec}{Canonical validation dataset specification.}
}
\value{
`TRUE`, invisibly.
}
\description{
Validate a canonical validation dataset specification
}
