% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/canonical_validation_suite.R
\name{validate_canonical_validation_suite}
\alias{validate_canonical_validation_suite}
\title{Validate a canonical validation suite}
\usage{
validate_canonical_validation_suite(suite)
}
\arguments{
\item{suite}{Canonical validation suite.}
}
\value{
`suite`, invisibly.
}
\description{
Validate a canonical validation suite
}
