% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzzz_manuscript_csl_integration.R
\name{validate_citation_profile}
\alias{validate_citation_profile}
\title{Validate a manuscript citation profile}
\usage{
validate_citation_profile(x)
}
\arguments{
\item{x}{A `PopgenVCFCitationProfile`.}
}
\value{
`TRUE` invisibly, or an error.
}
\description{
Validate a manuscript citation profile
}
