% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/continuous_release_benchmarking.R
\name{validate_continuous_benchmark_comparison}
\alias{validate_continuous_benchmark_comparison}
\title{Validate a continuous benchmark comparison}
\usage{
validate_continuous_benchmark_comparison(comparison)
}
\arguments{
\item{comparison}{Benchmark comparison.}
}
\value{
`comparison`, invisibly.
}
\description{
Validate a continuous benchmark comparison
}
