% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/continuous_release_benchmarking.R
\name{validate_continuous_benchmark_observation}
\alias{validate_continuous_benchmark_observation}
\title{Validate a continuous benchmark observation}
\usage{
validate_continuous_benchmark_observation(observation)
}
\arguments{
\item{observation}{Observation object.}
}
\value{
`observation`, invisibly.
}
\description{
Validate a continuous benchmark observation
}
