% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/execution-cancellation.R
\name{validate_execution_cancellation_token}
\alias{validate_execution_cancellation_token}
\title{Validate an execution cancellation token}
\usage{
validate_execution_cancellation_token(token)
}
\arguments{
\item{token}{A cancellation token.}
}
\value{
`token`, invisibly.
}
\description{
Validate an execution cancellation token
}
