% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/external-process.R
\name{validate_external_command}
\alias{validate_external_command}
\title{Validate an external command specification}
\usage{
validate_external_command(command)
}
\arguments{
\item{command}{A `PopgenVCFExternalCommand` object.}
}
\value{
`command`, invisibly.
}
\description{
Validate an external command specification
}
