% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/external_process_async.R
\name{validate_external_process_handle}
\alias{validate_external_process_handle}
\title{Validate an asynchronous external-process handle}
\usage{
validate_external_process_handle(handle)
}
\arguments{
\item{handle}{Object to validate.}
}
\value{
`handle`, invisibly.
}
\description{
Validate an asynchronous external-process handle
}
