% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/external-process-result-persistence.R
\name{validate_external_process_result}
\alias{validate_external_process_result}
\title{Validate an external process result}
\usage{
validate_external_process_result(result)
}
\arguments{
\item{result}{A `PopgenVCFExternalProcessResult` object.}
}
\value{
`result`, invisibly.
}
\description{
Validate an external process result
}
