% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/external_reference_validation.R
\name{validate_external_reference_spec}
\alias{validate_external_reference_spec}
\title{Validate an external-reference specification}
\usage{
validate_external_reference_spec(x)
}
\arguments{
\item{x}{A `PopgenVCFExternalReferenceSpec`.}
}
\value{
`x`, invisibly.
}
\description{
Validate an external-reference specification
}
