% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fair_export.R
\name{validate_fair_bundle}
\alias{validate_fair_bundle}
\title{Validate a FAIR research-object directory}
\usage{
validate_fair_bundle(directory)
}
\arguments{
\item{directory}{FAIR directory.}
}
\value{
`TRUE`, or an error.
}
\description{
Validate a FAIR research-object directory
}
