% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fair_export.R
\name{validate_fair_metadata}
\alias{validate_fair_metadata}
\title{Validate FAIR metadata}
\usage{
validate_fair_metadata(x)
}
\arguments{
\item{x}{A `PopgenVCFFAIRMetadata`.}
}
\value{
`x`, invisibly.
}
\description{
Validate FAIR metadata
}
