% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/golden_outputs.R
\name{validate_golden_entry}
\alias{validate_golden_entry}
\title{Validate a golden entry}
\usage{
validate_golden_entry(x)
}
\arguments{
\item{x}{A golden entry.}
}
\value{
`x`, invisibly.
}
\description{
Validate a golden entry
}
