% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzzzzzzzzz_graphical_abstract.R
\name{validate_graphical_abstract}
\alias{validate_graphical_abstract}
\title{Validate a graphical abstract specification}
\usage{
validate_graphical_abstract(x, strict = TRUE)
}
\arguments{
\item{x}{A `PopgenVCFGraphicalAbstract` object.}

\item{strict}{Whether missing author-supplied content is an error.}
}
\value{
`TRUE` invisibly.
}
\description{
Validate a graphical abstract specification
}
