% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzzzzzzz_journal_submission_profiles.R
\name{validate_journal_profile}
\alias{validate_journal_profile}
\title{Validate a journal submission profile}
\usage{
validate_journal_profile(x)
}
\arguments{
\item{x}{A `PopgenVCFJournalProfile` or written profile directory.}
}
\value{
`TRUE` invisibly.
}
\description{
Validate a journal submission profile
}
