% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manuscript_generation.R
\name{validate_manuscript}
\alias{validate_manuscript}
\title{Validate a canonical manuscript specification}
\usage{
validate_manuscript(x)
}
\arguments{
\item{x}{A `PopgenVCFManuscript` object or generated manuscript directory.}
}
\value{
`TRUE` invisibly, or an error.
}
\description{
Validate a canonical manuscript specification
}
