% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzzzzz_manuscript_jats.R
\name{validate_manuscript_jats}
\alias{validate_manuscript_jats}
\title{Validate generated JATS XML output}
\usage{
validate_manuscript_jats(x)
}
\arguments{
\item{x}{JATS text, a JATS file, or a `PopgenVCFJATSRecord`.}
}
\value{
`TRUE` invisibly, or an error.
}
\description{
Validate generated JATS XML output
}
