% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manuscript_revision.R
\name{validate_manuscript_revision}
\alias{validate_manuscript_revision}
\title{Validate a manuscript revision record}
\usage{
validate_manuscript_revision(x)
}
\arguments{
\item{x}{A `PopgenVCFManuscriptRevision`.}
}
\value{
`TRUE` invisibly.
}
\description{
Validate a manuscript revision record
}
