% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manuscript_revision.R
\name{validate_manuscript_revision_diff_bundle}
\alias{validate_manuscript_revision_diff_bundle}
\title{Validate a written manuscript revision diff bundle}
\usage{
validate_manuscript_revision_diff_bundle(directory)
}
\arguments{
\item{directory}{Written bundle directory.}
}
\value{
`TRUE` invisibly.
}
\description{
Validate a written manuscript revision diff bundle
}
