% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projects.R
\name{validate_popgenvcf_project}
\alias{validate_popgenvcf_project}
\title{Validate a reproducible project}
\usage{
validate_popgenvcf_project(x)
}
\arguments{
\item{x}{A `PopgenVCFProject`.}
}
\value{
`x`, invisibly.
}
\description{
Validate a reproducible project
}
