% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_engine.R
\name{validate_population_genomics_report_plan}
\alias{validate_population_genomics_report_plan}
\title{Validate a population-genomics report plan}
\usage{
validate_population_genomics_report_plan(x)
}
\arguments{
\item{x}{A `PopgenVCFReportPlan`.}
}
\value{
`x`, invisibly.
}
\description{
Validate a population-genomics report plan
}
