% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/provenance_dag.R
\name{validate_provenance_dag}
\alias{validate_provenance_dag}
\title{Validate a provenance DAG}
\usage{
validate_provenance_dag(dag)
}
\arguments{
\item{dag}{A provenance DAG.}
}
\value{
`dag`, invisibly.
}
\description{
Validate a provenance DAG
}
