% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/publication_amova_outputs.R
\name{validate_publication_amova_output}
\alias{validate_publication_amova_output}
\title{Validate a publication AMOVA output}
\usage{
validate_publication_amova_output(output, spec)
}
\arguments{
\item{output}{An AMOVA publication output.}

\item{spec}{Originating specification.}
}
\value{
`TRUE`, invisibly.
}
\description{
Validate a publication AMOVA output
}
