% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/publication_ancestry_outputs.R
\name{validate_publication_ancestry_output}
\alias{validate_publication_ancestry_output}
\title{Validate a publication ancestry output}
\usage{
validate_publication_ancestry_output(output, spec)
}
\arguments{
\item{output}{An ancestry publication output.}

\item{spec}{Originating specification.}
}
\value{
`TRUE`, invisibly.
}
\description{
Validate a publication ancestry output
}
