% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/publication_ancestry_outputs.R
\name{validate_publication_ancestry_spec}
\alias{validate_publication_ancestry_spec}
\title{Validate a publication ancestry specification}
\usage{
validate_publication_ancestry_spec(spec)
}
\arguments{
\item{spec}{An ancestry publication specification.}
}
\value{
`TRUE`, invisibly.
}
\description{
Validate a publication ancestry specification
}
