% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/publication_dapc_outputs.R
\name{validate_publication_dapc_output}
\alias{validate_publication_dapc_output}
\title{Validate a publication DAPC output}
\usage{
validate_publication_dapc_output(output, spec)
}
\arguments{
\item{output}{A publication DAPC output.}

\item{spec}{Originating specification.}
}
\value{
`TRUE`, invisibly.
}
\description{
Validate a publication DAPC output
}
