% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/publication_dapc_outputs.R
\name{validate_publication_dapc_spec}
\alias{validate_publication_dapc_spec}
\title{Validate a publication DAPC specification}
\usage{
validate_publication_dapc_spec(spec)
}
\arguments{
\item{spec}{A publication DAPC specification.}
}
\value{
`TRUE`, invisibly.
}
\description{
Validate a publication DAPC specification
}
