% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/publication_diversity_outputs.R
\name{validate_publication_diversity_output}
\alias{validate_publication_diversity_output}
\title{Validate a publication diversity output}
\usage{
validate_publication_diversity_output(output, spec)
}
\arguments{
\item{output}{A diversity publication output.}

\item{spec}{Originating specification.}
}
\value{
`TRUE`, invisibly.
}
\description{
Validate a publication diversity output
}
