% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/publication_diversity_outputs.R
\name{validate_publication_diversity_spec}
\alias{validate_publication_diversity_spec}
\title{Validate a publication diversity specification}
\usage{
validate_publication_diversity_spec(spec)
}
\arguments{
\item{spec}{A diversity publication specification.}
}
\value{
`TRUE`, invisibly.
}
\description{
Validate a publication diversity specification
}
