% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzzzzzz_publication_figure_styles.R
\name{validate_publication_figure_style_profile}
\alias{validate_publication_figure_style_profile}
\title{Validate a publication figure-style profile}
\usage{
validate_publication_figure_style_profile(profile)
}
\arguments{
\item{profile}{A figure-style profile.}
}
\value{
`TRUE`, invisibly.
}
\description{
Validate a publication figure-style profile
}
