% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/publication_fst_outputs.R
\name{validate_publication_fst_output}
\alias{validate_publication_fst_output}
\title{Validate a publication FST output}
\usage{
validate_publication_fst_output(output, spec)
}
\arguments{
\item{output}{A publication FST output.}

\item{spec}{Originating specification.}
}
\value{
`TRUE`, invisibly.
}
\description{
Validate a publication FST output
}
