% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/publication_fst_outputs.R
\name{validate_publication_fst_spec}
\alias{validate_publication_fst_spec}
\title{Validate a publication FST specification}
\usage{
validate_publication_fst_spec(spec)
}
\arguments{
\item{spec}{A publication FST specification.}
}
\value{
`TRUE`, invisibly.
}
\description{
Validate a publication FST specification
}
