% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/publication_ibd_outputs.R
\name{validate_publication_ibd_output}
\alias{validate_publication_ibd_output}
\title{Validate a publication IBD output}
\usage{
validate_publication_ibd_output(output, spec)
}
\arguments{
\item{output}{An IBD publication output.}

\item{spec}{Originating specification.}
}
\value{
`TRUE`, invisibly.
}
\description{
Validate a publication IBD output
}
