% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/publication_ibd_outputs.R
\name{validate_publication_ibd_spec}
\alias{validate_publication_ibd_spec}
\title{Validate a publication IBD specification}
\usage{
validate_publication_ibd_spec(spec)
}
\arguments{
\item{spec}{An IBD publication specification.}
}
\value{
`TRUE`, invisibly.
}
\description{
Validate a publication IBD specification
}
