% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzzzz_publication_layout_profiles.R
\name{validate_publication_layout_binding}
\alias{validate_publication_layout_binding}
\title{Validate a publication layout binding}
\usage{
validate_publication_layout_binding(binding, spec, profile)
}
\arguments{
\item{binding}{A layout binding.}

\item{spec}{Originating report specification.}

\item{profile}{Originating layout profile.}
}
\value{
`TRUE`, invisibly.
}
\description{
Validate a publication layout binding
}
