% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzzzz_publication_layout_profiles.R
\name{validate_publication_layout_profile}
\alias{validate_publication_layout_profile}
\title{Validate a publication layout profile}
\usage{
validate_publication_layout_profile(profile, journal_profile = NULL)
}
\arguments{
\item{profile}{A publication layout profile.}

\item{journal_profile}{Optional originating journal profile.}
}
\value{
`TRUE`, invisibly.
}
\description{
Validate a publication layout profile
}
