% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/publication_ordination_outputs.R
\name{validate_publication_ordination_output}
\alias{validate_publication_ordination_output}
\title{Validate a publication ordination output}
\usage{
validate_publication_ordination_output(output, spec)
}
\arguments{
\item{output}{A publication ordination output.}

\item{spec}{Originating specification.}
}
\value{
`TRUE`, invisibly.
}
\description{
Validate a publication ordination output
}
