% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/publication_ordination_outputs.R
\name{validate_publication_ordination_spec}
\alias{validate_publication_ordination_spec}
\title{Validate a publication ordination specification}
\usage{
validate_publication_ordination_spec(spec)
}
\arguments{
\item{spec}{A publication ordination specification.}
}
\value{
`TRUE`, invisibly.
}
\description{
Validate a publication ordination specification
}
