% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zz_publication_report_execution.R
\name{validate_publication_report_execution}
\alias{validate_publication_report_execution}
\title{Validate a publication report execution record}
\usage{
validate_publication_report_execution(execution, plan, output_dir)
}
\arguments{
\item{execution}{An execution record.}

\item{plan}{The originating plan.}

\item{output_dir}{Destination directory used for execution.}
}
\value{
`TRUE`, invisibly.
}
\description{
Validate a publication report execution record
}
