% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/publication_report_rendering.R
\name{validate_publication_report_plan}
\alias{validate_publication_report_plan}
\title{Validate a publication report rendering plan}
\usage{
validate_publication_report_plan(plan, manuscript, spec)
}
\arguments{
\item{plan}{A publication report rendering plan.}

\item{manuscript}{The originating manuscript.}

\item{spec}{The originating rendering specification.}
}
\value{
`TRUE`, invisibly.
}
\description{
Validate a publication report rendering plan
}
