% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zz_publication_report_execution.R
\name{validate_publication_report_renderer}
\alias{validate_publication_report_renderer}
\title{Validate a publication report renderer adapter}
\usage{
validate_publication_report_renderer(renderer)
}
\arguments{
\item{renderer}{A renderer adapter.}
}
\value{
`TRUE`, invisibly.
}
\description{
Validate a publication report renderer adapter
}
