% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/publication_report_rendering.R
\name{validate_publication_report_spec}
\alias{validate_publication_report_spec}
\title{Validate a publication report specification}
\usage{
validate_publication_report_spec(spec)
}
\arguments{
\item{spec}{A publication report specification.}
}
\value{
`TRUE`, invisibly.
}
\description{
Validate a publication report specification
}
