% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/publication_spatial_genetics.R
\name{validate_publication_spatial_output}
\alias{validate_publication_spatial_output}
\title{Validate a publication spatial-genetics output}
\usage{
validate_publication_spatial_output(output, spec)
}
\arguments{
\item{output}{A spatial-genetics publication output.}

\item{spec}{Originating specification.}
}
\value{
`TRUE`, invisibly.
}
\description{
Validate a publication spatial-genetics output
}
