% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzzzzzzz_publication_submission_packages.R
\name{validate_publication_submission_package_manifest}
\alias{validate_publication_submission_package_manifest}
\title{Validate a submission-package manifest}
\usage{
validate_publication_submission_package_manifest(
  manifest,
  spec,
  supplementary_index
)
}
\arguments{
\item{manifest}{A package manifest.}

\item{spec}{Originating package specification.}

\item{supplementary_index}{Originating supplementary index.}
}
\value{
`TRUE`, invisibly.
}
\description{
Validate a submission-package manifest
}
