% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzzzzzzz_publication_submission_packages.R
\name{validate_publication_submission_package_spec}
\alias{validate_publication_submission_package_spec}
\title{Validate a submission-package specification}
\usage{
validate_publication_submission_package_spec(
  spec,
  journal_profile,
  layout_profile,
  figure_style,
  report_spec
)
}
\arguments{
\item{spec}{A submission-package specification.}

\item{journal_profile}{Originating journal profile.}

\item{layout_profile}{Originating layout profile.}

\item{figure_style}{Originating figure-style profile.}

\item{report_spec}{Originating report specification.}
}
\value{
`TRUE`, invisibly.
}
\description{
Validate a submission-package specification
}
