% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzzzzzzz_publication_submission_packages.R
\name{validate_publication_supplementary_index}
\alias{validate_publication_supplementary_index}
\title{Validate a supplementary-material index}
\usage{
validate_publication_supplementary_index(index)
}
\arguments{
\item{index}{A supplementary-material index.}
}
\value{
`TRUE`, invisibly.
}
\description{
Validate a supplementary-material index
}
