% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regression_archive.R
\name{validate_release_benchmark_record}
\alias{validate_release_benchmark_record}
\title{Validate a release benchmark record}
\usage{
validate_release_benchmark_record(x)
}
\arguments{
\item{x}{A release benchmark record.}
}
\value{
`x`, invisibly.
}
\description{
Validate a release benchmark record
}
