% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/continuous_release_benchmarking.R
\name{validate_release_performance_budget}
\alias{validate_release_performance_budget}
\title{Validate a release performance budget}
\usage{
validate_release_performance_budget(budget)
}
\arguments{
\item{budget}{Release performance budget.}
}
\value{
`budget`, invisibly.
}
\description{
Validate a release performance budget
}
