% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzzzzzzzz_reporting_checklists.R
\name{validate_reporting_checklist}
\alias{validate_reporting_checklist}
\title{Validate a reporting checklist or written checklist directory}
\usage{
validate_reporting_checklist(x)
}
\arguments{
\item{x}{A `PopgenVCFReportingChecklist` or written checklist directory.}
}
\value{
`TRUE` invisibly.
}
\description{
Validate a reporting checklist or written checklist directory
}
