% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_identity.R
\name{validate_sample_identity}
\alias{validate_sample_identity}
\title{Validate a canonical sample identity table}
\usage{
validate_sample_identity(x)
}
\arguments{
\item{x}{A sample identity table.}
}
\value{
`x`, invisibly.
}
\description{
Validate a canonical sample identity table
}
