% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scheduler-metadata-integrity.R
\name{validate_scheduler_metadata}
\alias{validate_scheduler_metadata}
\title{Validate scheduler metadata}
\usage{
validate_scheduler_metadata(metadata)
}
\arguments{
\item{metadata}{Scheduler metadata.}
}
\value{
`metadata`, invisibly.
}
\description{
Validate scheduler metadata
}
