% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzzzzzz_manuscript_submission_package.R
\name{verify_submission_package}
\alias{verify_submission_package}
\title{Verify a manuscript submission package}
\usage{
verify_submission_package(path)
}
\arguments{
\item{path}{Submission `.tar.gz` archive.}
}
\value{
`TRUE` invisibly, or an error.
}
\description{
Verify a manuscript submission package
}
