% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/artifact_lineage.R
\name{write_artifact_lineage}
\alias{write_artifact_lineage}
\title{Export immutable artifact lineage}
\usage{
write_artifact_lineage(
  lineage,
  directory,
  formats = c("tsv", "json", "graphml", "dot")
)
}
\arguments{
\item{lineage}{An artifact lineage object.}

\item{directory}{Output directory.}

\item{formats}{Any of `tsv`, `json`, `graphml`, and `dot`.}
}
\value{
Named vector of written files.
}
\description{
Export immutable artifact lineage
}
