% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzzzzzzzzz_graphical_abstract.R
\name{write_graphical_abstract}
\alias{write_graphical_abstract}
\title{Write a graphical abstract specification bundle}
\usage{
write_graphical_abstract(
  graphical_abstract,
  directory,
  overwrite = FALSE,
  strict = FALSE
)
}
\arguments{
\item{graphical_abstract}{A `PopgenVCFGraphicalAbstract` object.}

\item{directory}{Parent output directory.}

\item{overwrite}{Whether an existing output directory may be replaced.}

\item{strict}{Whether incomplete author-supplied content is an error.}
}
\value{
The normalized output directory invisibly.
}
\description{
Write a graphical abstract specification bundle
}
