% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_publication.R
\name{write_publication_bibliography}
\alias{write_publication_bibliography}
\title{Write a BibTeX library for analyses present in a project}
\usage{
write_publication_bibliography(narratives, path)
}
\arguments{
\item{narratives}{Output from `publication_analysis_narratives()`.}

\item{path}{Destination `.bib` path.}
}
\value{
Normalized path, invisibly.
}
\description{
Write a BibTeX library for analyses present in a project
}
